+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MeasuringFourierTransforms"
+            android:name=".activities.MeasuringFourierTransforms"
             android:label="@string/title_activity_measuring_fourier_transforms"
             android:parentActivityName=".MainActivity"
             android:theme="@style/AppTheme.NoActionBar">
@@ -30,7 +30,7 @@
                 android:value="de.voicegym.voicegym.MainActivity" />
         </activity>
         <activity
-            android:name=".Activities.RecordActivity"
+            android:name=".activities.RecordActivity"
             android:configChanges="orientation|keyboardHidden|screenSize"
             android:label="@string/title_activity_record"
             android:theme="@style/FullscreenNoActionbar" />
