+++ app/src/main/java/de/voicegym/voicegym/fourierHelper/AsyncTransformer.kt
@@ -1,9 +1,9 @@
-package de.voicegym.voicegym.FourierHelper
+package de.voicegym.voicegym.fourierHelper
 
-import de.voicegym.voicegym.FourierHelper.RequestedResultType.FFT_AMPLITUDE
-import de.voicegym.voicegym.FourierHelper.RequestedResultType.FFT_AMPLITUDE_AND_PHASE
-import de.voicegym.voicegym.FourierHelper.RequestedResultType.FFT_COMPLEX_RESULT
-import de.voicegym.voicegym.FourierHelper.RequestedResultType.FFT_PHASE
+import de.voicegym.voicegym.fourierHelper.RequestedResultType.FFT_AMPLITUDE
+import de.voicegym.voicegym.fourierHelper.RequestedResultType.FFT_AMPLITUDE_AND_PHASE
+import de.voicegym.voicegym.fourierHelper.RequestedResultType.FFT_COMPLEX_RESULT
+import de.voicegym.voicegym.fourierHelper.RequestedResultType.FFT_PHASE
 import java.util.Observable
 import java.util.concurrent.ConcurrentLinkedQueue
 import kotlin.concurrent.thread
