+++ app/src/main/java/de/voicegym/voicegym/fourierHelper/PCMUtil.kt
@@ -1,4 +1,4 @@
-package de.voicegym.voicegym.FourierHelper
+package de.voicegym.voicegym.fourierHelper
 
 object PCMUtil {
     fun getDoubleArrayFromShortArray(normalization: Double, inputArray: ShortArray): DoubleArray {
