+++ app/src/main/java/de/voicegym/voicegym/views/util/GradientPicker.kt
@@ -1,4 +1,4 @@
-package de.voicegym.voicegym.Activities.InstrumentViews
+package de.voicegym.voicegym.views.util
 
 interface GradientPicker {
     fun pickColor(value: Double): Int
