+++ app/src/main/java/de/voicegym/voicegym/views/util/HotGradientColorPicker.kt
@@ -1,4 +1,4 @@
-package de.voicegym.voicegym.Activities.InstrumentViews
+package de.voicegym.voicegym.views.util
 
 import android.graphics.Color
 import android.graphics.Color.rgb
