+++ app/src/main/res/layout/activity_measuring_fourier_transforms.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MeasuringFourierTransforms">
+    tools:context=".activities.MeasuringFourierTransforms">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
