+++ app/src/main/res/layout/activity_record.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
 
-    tools:context=".Activities.RecordActivity">
+    tools:context=".activities.RecordActivity">
 
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/floatingActionButton"
@@ -25,7 +25,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <de.voicegym.voicegym.Activities.InstrumentViews.SpectrogramView
+        <de.voicegym.voicegym.views.SpectrogramView
             android:id="@+id/dummyView"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
